Deleted all files which are not used in the current approach: and renamed all *2 definitions back to normal. E. g. Graph2 -> Graph
