--- src/genericAlgorithm/basic_genetic_algorithm.py
@@ -1,125 +1,0 @@
-from random import choices, randint, randrange, random
-from typing import List, Optional, Callable, Tuple
-from tqdm import tqdm
-import logging
-
-Genome = List[int]
-Population = List[Genome]
-PopulateFunc = Callable[[], Population]
-FitnessFunc = Callable[[Genome], int]
-SelectionFunc = Callable[[Population, FitnessFunc], Tuple[Genome, Genome]]
-CrossoverFunc = Callable[[Genome, Genome], Tuple[Genome, Genome]]
-MutationFunc = Callable[[Genome], Genome]
-PrinterFunc = Callable[[Population, int, FitnessFunc], None]
-
-
-def generate_genome(length: int) -> Genome:
-    return choices([0, 1], k=length)
-
-
-def generate_population(size: int, genome_length: int) -> Population:
-    return [generate_genome(genome_length) for _ in range(size)]
-
-
-def single_point_crossover(a: Genome, b: Genome) -> Tuple[Genome, Genome]:
-    if len(a) != len(b):
-        raise ValueError("Genomes a and b must be of same length")
-
-    length = len(a)
-    if length < 2:
-        return a, b
-
-    p = randint(1, length - 1)
-    return a[0:p] + b[p:], b[0:p] + a[p:]
-
-
-def mutation(genome: Genome, num: int = 1, probability: float = 0.5) -> Genome:
-    for _ in range(num):
-        index = randrange(len(genome))
-        genome[index] = (
-            genome[index] if random() > probability else abs(genome[index] - 1)
-        )
-    return genome
-
-
-def population_fitness(population: Population, fitness_func: FitnessFunc) -> int:
-    return sum([fitness_func(genome) for genome in population])
-
-
-def selection_pair(population: Population, fitness_func: FitnessFunc) -> Population:
-    return choices(
-        population=population, weights=[fitness_func(gene) for gene in population], k=2
-    )
-
-
-def sort_population(population: Population, fitness_func: FitnessFunc) -> Population:
-    return sorted(population, key=fitness_func, reverse=True)
-
-
-def genome_to_string(genome: Genome) -> str:
-    return "".join(map(str, genome))
-
-
-def print_stats(population: Population, generation_id: int, fitness_func: FitnessFunc):
-    sorted_population = sort_population(population, fitness_func)
-    logging.debug(
-        "GENERATION {}: avg: {}\tBest: {}\tWorst: {}".format(
-            generation_id,
-            (population_fitness(population, fitness_func) / len(population)),
-            fitness_func(sorted_population[0]),
-            fitness_func(sorted_population[-1]),
-        )
-    )
-    # logging.info("GENERATION %02d: avg: %f\t\tBest: %f\t\tWorst: %f" % generation_id)
-    # logging.info("=============")
-    # logging.info("Population: [%s]" % ", ".join([genome_to_string(gene) for gene in population]))
-    # logging.info("Avg. Fitness: %f" % (population_fitness(population, fitness_func) / len(population)))
-    # logging.info("Best: (%f)" % fitness_func(sorted_population[0]))
-    # logging.info("Worst: (%f)" % fitness_func(sorted_population[-1]))
-    # logging.info("Worst: %s (%f)" % (genome_to_string(sorted_population[-1]),
-    #                           fitness_func(sorted_population[-1])))
-
-    # logging.info(
-    #     "Best: %s (%f)" % (genome_to_string(sorted_population[0]), fitness_func(sorted_population[0])))
-    # logging.info("Worst: %s (%f)" % (genome_to_string(sorted_population[-1]),
-    #                           fitness_func(sorted_population[-1])))
-    # logging.info("")
-
-    return sorted_population[0]
-
-
-def run_evolution(
-    populate_func: PopulateFunc,
-    fitness_func: FitnessFunc,
-    fitness_limit: int,
-    selection_func: SelectionFunc = selection_pair,
-    crossover_func: CrossoverFunc = single_point_crossover,
-    mutation_func: MutationFunc = mutation,
-    generation_limit: int = 100,
-    printer: Optional[PrinterFunc] = None,
-) -> Tuple[Population, int]:
-    population = populate_func()
-
-    for i in tqdm(range(generation_limit), mininterval=0.1, miniters=0):
-        population = sorted(
-            population, key=lambda genome: fitness_func(genome), reverse=True
-        )
-
-        if printer is not None:
-            printer(population, i, fitness_func)
-
-        if fitness_func(population[0]) >= fitness_limit:
-            break
-
-        next_generation = population[0:2]
-
-        for j in range(int(len(population) / 2) - 1):
-            parents = selection_func(population, fitness_func)
-            offspring_a, offspring_b = crossover_func(parents[0], parents[1])
-            offspring_a = mutation_func(offspring_a)
-            offspring_b = mutation_func(offspring_b)
-            next_generation += [offspring_a, offspring_b]
-
-        population = next_generation
-
-    return population, i
